handle file drop at page level for player
