+++ Screenbox.Core/ViewModels/PlayerPageViewModel.cs
@@ -13,9 +13,12 @@ using Screenbox.Core.Models;
 using Screenbox.Core.Playback;
 using Screenbox.Core.Services;
 using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
+using Windows.ApplicationModel.DataTransfer;
 using Windows.Media;
 using Windows.Media.Playback;
+using Windows.Storage;
 using Windows.System;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
@@ -179,6 +182,45 @@ namespace Screenbox.Core.ViewModels
         public void Receive(OverrideControlsHideDelayMessage message)
         {
             OverrideControlsDelayHide(message.Delay);
+        }
+
+        public async void OnDrop(object sender, DragEventArgs e)
+        {
+            if (_mediaPlayer == null) return;
+            try
+            {
+                if (e.DataView.Contains(StandardDataFormats.StorageItems))
+                {
+                    IReadOnlyList<IStorageItem>? items = await e.DataView.GetStorageItemsAsync();
+                    if (items.Count > 0)
+                    {
+                        if (items.Count == 1 && items[0] is StorageFile { FileType: ".srt" or ".ass" } file)
+                        {
+                            _mediaPlayer.AddSubtitle(file);
+                            Messenger.Send(new SubtitleAddedNotificationMessage(file));
+                        }
+                        else
+                        {
+                            Messenger.Send(new PlayFilesWithNeighborsMessage(items, null));
+                        }
+
+                        return;
+                    }
+                }
+
+                if (e.DataView.Contains(StandardDataFormats.WebLink))
+                {
+                    Uri? uri = await e.DataView.GetWebLinkAsync();
+                    if (uri.IsFile)
+                    {
+                        Messenger.Send(new PlayMediaMessage(uri));
+                    }
+                }
+            }
+            catch (Exception exception)
+            {
+                Messenger.Send(new MediaLoadFailedNotificationMessage(exception.Message, string.Empty));
+            }
         }
 
         public bool OnPlayerClick()
