+++ Screenbox/Controls/PlayerElement.xaml
@@ -20,11 +20,8 @@
             Margin="{x:Bind ButtonMargin, Mode=OneWay}"
             HorizontalAlignment="Stretch"
             VerticalAlignment="Stretch"
-            AllowDrop="True"
             AllowFocusOnInteraction="False"
             DoubleTapped="VideoViewButton_OnDoubleTapped"
-            DragOver="VideoViewButton_OnDragOver"
-            Drop="{x:Bind ViewModel.OnDrop}"
             IsEnabled="{x:Bind IsEnabled, Mode=OneWay}"
             ManipulationCompleted="{x:Bind ViewModel.VideoView_ManipulationCompleted}"
             ManipulationDelta="{x:Bind ViewModel.VideoView_ManipulationDelta}"
