+++ Screenbox/Pages/PlayerPage.xaml
@@ -21,6 +21,9 @@
     toolkit:TitleBarExtensions.ButtonBackgroundColor="Transparent"
     toolkit:TitleBarExtensions.ButtonInactiveBackgroundColor="Transparent"
     toolkit:TitleBarExtensions.InactiveBackgroundColor="Transparent"
+    AllowDrop="True"
+    DragOver="OnDragOver"
+    Drop="{x:Bind ViewModel.OnDrop}"
     Loaded="OnLoaded"
     Loading="OnLoading"
     mc:Ignorable="d">
